+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - Add: sorting uses `slices` package.
 
+## [v1.8.0] - 2023-10-11 Wed
+
+- Add [#255]: normalize stemmed canonical of `Aus bus bus` to `Aus bus`.
+
 ## [v1.7.5] - 2023-09-26 Tue
 
 - Add: CSV and TSV files provide now verbatim authorship instead of normalized